cleaned up comments from context files
