+++ context/AuthContext.tsx
@@ -90,8 +90,6 @@ export const AuthContextProvider = ({
   };
 
   const signupHandler = async (user: User) => {
-    //@ts-ignore
-    user.passwordConfirm = user.password;
     try {
       const res = await axios.post("/api/auth/signup", user, {
         headers: {
@@ -106,7 +104,6 @@ export const AuthContextProvider = ({
         token: res.data.token,
       });
     } catch (err: any) {
-      console.log("Caught");
       if (err.response) {
         throw err.response.data.error;
       }
