+++ pages/index.js
@@ -4,11 +4,11 @@ import Image from "next/image";
 import styles from "../styles/Home.module.css";
 
 function Home() {
-  return (
-    <div className={styles.container}>
-     <a>NavBar</a>
-    </div>
-  );
+	return (
+		<div className={styles.container}>
+      Home Page
+		</div>
+	);
 }
 
 export default Home;
