+++ pages/register.js
@@ -1,119 +1,117 @@
 import Link from "next/link";
-import React, {useContext, useEffect, useState} from "react";
+import React, { useContext, useState } from "react";
 
 import AuthContext from "../context/AuthContext";
 import styles from "../styles/Login.module.css";
 
 export default function Register() {
-  // states
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
+	// states
+	const [name, setName] = useState("");
+	const [email, setEmail] = useState("");
+	const [password, setPassword] = useState("");
+	const [confirmPassword, setConfirmPassword] = useState("");
 
-  const [passwordWarning, setPasswordWarning] = useState("");
-  const [diffPasswords, setDiffPasswords] = useState(false);
-  const {signup, user} = useContext(AuthContext);
-  useEffect(() => {
-    console.log(user);
-  }, [user]);
-  // event handlers
-  const handleName = (event) => {
-    setName(event.target.value);
-  };
+	const [passwordWarning, setPasswordWarning] = useState("");
+	const [diffPasswords, setDiffPasswords] = useState(false);
 
-  const handleEmail = (event) => {
-    setEmail(event.target.value);
-  };
+  const { signup } = useContext(AuthContext);
+	
+	// event handlers
+	const handleName = (event) => {
+		setName(event.target.value);
+	};
 
-  const handlePassword = (event) => {
-    setPassword(event.target.value);
-  };
+	const handleEmail = (event) => {
+		setEmail(event.target.value);
+	};
 
-  const handleConfirmPassword = (event) => {
-    setConfirmPassword(event.target.value);
-  };
+	const handlePassword = (event) => {
+		setPassword(event.target.value);
+	};
 
-  function submitForm(event) {
-    event.preventDefault();
+	const handleConfirmPassword = (event) => {
+		setConfirmPassword(event.target.value);
+	};
 
-    if (password === confirmPassword) {
-      // create a new user and send to the back end
+	function submitForm(event) {
+		event.preventDefault();
 
-      const createUser = {
-        name: name,
-        email: email,
-        password: password,
-      };
-      signup(createUser);
-      // TODO: check if the user details are saved in the object
+		if (password === confirmPassword) {
+			// create a new user and send to the back end
 
-      // TODO: send the user info to the backend
-    } else {
-      // create an error message (inconsistent password) and displays it
-      console.log(diffPasswords);
+			const newUser = {
+				name: name,
+				email: email,
+				password: password,
+        passwordConfirm: confirmPassword
 
-      setDiffPasswords(true);
-      setPasswordWarning("Passwords do not match!");
-    }
+			};
 
-    return;
-  }
+			signup(newUser);
+		} else {
+			// create an error message (inconsistent password) and displays it
 
-  return (
-    <div className={styles.loginBox}>
-      <h4>Sign up! Enter your name, email and password.</h4>
-      <br />
+			setDiffPasswords(true);
+			setPasswordWarning("Passwords do not match!");
+		}
 
-      {diffPasswords && <div>{passwordWarning}</div>}
+		return;
+	}
 
-      <form onSubmit={submitForm}>
-        <label>Name:</label>
-        <input
-          type="text"
-          id="name"
-          name="name"
-          value={name}
-          onChange={handleName}
-          required
-        />
-        <br />
-        <label>Email:</label>
-        <input
-          type="text"
-          id="email"
-          name="email"
-          value={email}
-          onChange={handleEmail}
-          required
-        />
-        <br />
-        <label>Password:</label>
-        <input
-          type="text"
-          id="pw"
-          name="pw"
-          value={password}
-          onChange={handlePassword}
-          required
-        />
-        <br />
-        <label>Confirm Password:</label>
-        <input
-          type="text"
-          id="repeat_pw"
-          name="repeat_pw"
-          value={confirmPassword}
-          onChange={handleConfirmPassword}
-          required
-        />
-        <br />
-        <button type="submit">Submit</button>
-        <br />
-        <Link href="/login">
-          <a>Already Have An Account?</a>
-        </Link>
-      </form>
-    </div>
-  );
+	return (
+		<div className={styles.loginBox}>
+			<h4>Sign up! Enter your name, email and password.</h4>
+			<br />
+
+			{diffPasswords && <div>{passwordWarning}</div>}
+
+			<form onSubmit={submitForm}>
+				<label>Name:</label>
+				<input
+					type="text"
+					id="name"
+					name="name"
+					value={name}
+					onChange={handleName}
+					required
+				/>
+				<br />
+				<label>Email:</label>
+				<input
+					type="text"
+					id="email"
+					name="email"
+					value={email}
+					onChange={handleEmail}
+					required
+				/>
+				<br />
+				<label>Password:</label>
+				<input
+					type="text"
+					id="pw"
+					name="pw"
+					value={password}
+					onChange={handlePassword}
+					required
+				/>
+				<br />
+				<label>Confirm Password:</label>
+				<input
+					type="text"
+					id="repeat_pw"
+					name="repeat_pw"
+					value={confirmPassword}
+					onChange={handleConfirmPassword}
+					required
+				/>
+				<br />
+				<button type="submit">Submit</button>
+				<br />
+				<Link href="/login">
+					<a>Already Have An Account?</a>
+				</Link>
+			</form>
+		</div>
+	);
 }
